+++ src/runtimes/node/bundlers/zisi/src_files.js
@@ -3,7 +3,7 @@ const { dirname, basename, normalize } = require('path')
 const { not: notJunk } = require('junk')
 const precinct = require('precinct')
 
-const { filterExcludedPaths, getPathsOfIncludedFiles } = require('../../src_files')
+const { filterExcludedPaths, getPathsOfIncludedFiles } = require('../../utils/included_files')
 const { getPackageJson } = require('../../utils/package_json')
 const { getNewCache } = require('../../utils/traversal_cache')
